Make blocks final
