+++ src/Block/FeatureMediaBlockService.php
@@ -23,11 +23,9 @@ use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
 /**
- * @final since sonata-project/media-bundle 3.21.0
- *
  * @author Thomas Rabaix <thomas.rabaix@sonata-project.org>
  */
-class FeatureMediaBlockService extends MediaBlockService
+final class FeatureMediaBlockService extends BaseMediaBlockService
 {
     public function configureSettings(OptionsResolver $resolver): void
     {
